move sample to text
